+++ src/components/calcite-card/calcite-card.e2e.ts
@@ -1,67 +1,43 @@
 import { newE2EPage } from "@stencil/core/testing";
-import { defaults, renders } from "../../tests/commonTests";
+import { renders } from "../../tests/commonTests";
 import { CSS } from "./resources";
 
 describe("calcite-card", () => {
   it("renders", async () => renders("calcite-card"));
 
-  it("has property defaults", async () => {
-    defaults("calcite-card", [
-      {
-        propertyName: "disabled",
-        defaultValue: false
-      },
-      {
-        propertyName: "loading",
-        defaultValue: false,
-      },
-      {
-        propertyName: "respectImageHeight",
-        defaultValue: false
-      },
-      {
-        propertyName: "selected",
-        defaultValue:false
-      },
-      {
-        propertyName: "theme",
-        defaultValue: "light"
-      }
-    ])
-  });
-
-  it("renders default props when none are provided", async () => {
-    const page = await newE2EPage();
-    await page.setContent(`
-    <calcite-card>
-      <img slot="thumbnail" src="https://via.placeholder.com/350x150.png" />
-      <div slot="header">heyo</div>
-      <calcite-button slot="button-action">button</calcite-button>
-      <div slot="footer-leading">footer1</div>
-      <div slot="footer-trailing">footer2</div>
-    </calcite-card>
-    `);
-    const card = await page.find("calcite-card");
-    expect(card).toHaveClass("hydrated");
-    expect(card).toEqualAttribute("theme", "light");
-  });
-
-  it("renders default props when invalid props are provided", async () => {
-    const page = await newE2EPage();
-    await page.setContent(`
-    <calcite-card theme="jungle">
-    </calcite-card>
-    `);
-
-    const card = await page.find("calcite-card");
-    expect(card).toEqualAttribute("theme", "light");
-  })
-
-  it("should have an thumbnail container", async() => {
+  it("renders with default props if none are provided", async () => {
     const page = await newE2EPage();
     await page.setContent(`
       <calcite-card>
-      <img slot="thumbnail" src="https://via.placeholder.com/350x150.png" />
+        <img slot="thumbnail" src="https://via.placeholder.com/350x150.png" />
+      </calcite-card>`);
+
+    const element = await page.find("calcite-card");
+    expect(element).not.toHaveAttribute("disabled");
+    expect(element).not.toHaveAttribute("loading");
+    expect(element).not.toHaveAttribute("selectable");
+    expect(element).not.toHaveAttribute("selected");
+  });
+
+  it("renders with requsted props", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+      <calcite-card loading selectable selected disabled>
+        <img slot="thumbnail" src="https://via.placeholder.com/350x150.png" />
+      </calcite-card>`);
+
+    const element = await page.find("calcite-card");
+    expect(element).toHaveAttribute("disabled");
+    expect(element).toHaveAttribute("loading");
+    expect(element).toHaveAttribute("selectable");
+    expect(element).toHaveAttribute("selected");
+  });
+
+  it("should have a thumbnail container", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+      <calcite-card>
+        <img slot="thumbnail" src="https://via.placeholder.com/350x150.png" />
       </calcite-card>
     `);
 
@@ -69,10 +45,10 @@ describe("calcite-card", () => {
       `calcite-card >>> .${CSS.thumbnailWrapper}`
     );
 
-    expect( await thumbContainer.isVisible()).toBe(true);
+    expect(thumbContainer).not.toBeNull();
   });
 
-  it("should render a checkbox if selectable", async() => {
+  it("should render a checkbox if selectable", async () => {
     const page = await newE2EPage();
     await page.setContent(`
       <calcite-card selectable>
@@ -81,10 +57,9 @@ describe("calcite-card", () => {
     `);
 
     const thumbContainer = await page.find(
-      `calcite-card >>> .${CSS.thumbnailWrapper}`
+      `calcite-card >>> .${CSS.checkboxWrapper}`
     );
 
-    expect( await thumbContainer.isVisible()).toBe(true);
+    expect(thumbContainer).not.toBeNull();
   });
-
-})
+});