+++ src/components/calcite-icon/calcite-icon.e2e.ts
@@ -13,7 +13,6 @@ describe("calcite-icon", () => {
     defaults("calcite-icon", [
       { propertyName: "filled", defaultValue: false },
       { propertyName: "mirrored", defaultValue: false },
-      { propertyName: "theme", defaultValue: "light" },
       { propertyName: "scale", defaultValue: "m" }
     ]));
 
@@ -21,12 +20,13 @@ describe("calcite-icon", () => {
     reflects("calcite-icon", [
       { propertyName: "filled", value: true },
       { propertyName: "mirrored", value: true },
-      { propertyName: "theme", value: "light" },
       { propertyName: "scale", value: "m" }
     ]));
 
   it("is accessible", async () =>
-    accessible(`<calcite-icon icon="a-z" text-label="sort options"></calcite-icon>`));
+    accessible(
+      `<calcite-icon icon="a-z" text-label="sort options"></calcite-icon>`
+    ));
 
   it("mirrors icon when enabled and in RTL", async () => {
     const page = await newE2EPage();
@@ -53,7 +53,6 @@ describe("calcite-icon", () => {
       const path = await page.find(`calcite-icon >>> path`);
 
       expect(await path.getAttribute("d")).toBeTruthy();
-
     });
 
     it("supports both camelcase and kebab case for icon name", async () => {
@@ -70,6 +69,18 @@ describe("calcite-icon", () => {
       expect(await path.getAttribute("d")).toBe(iconPathData);
     });
 
+    it("it does not render SVG for invalid icons", async () => {
+      const page = await newE2EPage();
+      await page.setContent(
+        `<calcite-icon icon="this-does-not-exist"></calcite-icon>`
+      );
+      await page.waitForChanges();
+
+      let svg = await page.find(`calcite-icon >>> svg`);
+
+      expect(svg).toBeNull();
+    });
+
     it("loads icon when it's close to viewport", async () => {
       const page = await newE2EPage();
       await page.setContent(
@@ -77,13 +88,17 @@ describe("calcite-icon", () => {
       );
       await page.waitForChanges();
 
-      const icon = await page.find(`calcite-icon`);
-      let path = await page.find(`calcite-icon >>> path`);
+      const iconPathSelector = `calcite-icon >>> path`;
 
-      expect(await path.getAttribute("d")).toBeNull();
+      const icon = await page.find(`calcite-icon`);
+      let path = await page.find(iconPathSelector);
+
+      expect(path).toBeNull();
 
       icon.setProperty("style", null);
       await page.waitForChanges();
+
+      path = await page.find(iconPathSelector);
 
       expect(await path.getAttribute("d")).toBeTruthy();
     });