+++ src/components/calcite-switch/calcite-switch.tsx
@@ -11,7 +11,6 @@ import {
   Build
 } from "@stencil/core";
 import { SPACE, ENTER } from "../../utils/keys";
-import { getElementDir } from "../../utils/dom";
 
 @Component({
   tag: "calcite-switch",
@@ -37,33 +36,31 @@ export class CalciteSwitch {
   @Prop({ reflect: true, mutable: true }) scale: "s" | "m" | "l" = "m";
 
   /** The component's theme. */
-  @Prop({ reflect: true, mutable: true }) theme: "light" | "dark" = "light";
+  @Prop({ reflect: true, mutable: true }) theme: "light" | "dark";
 
   @Event() calciteSwitchChange: EventEmitter;
+  @Event() change: EventEmitter;
 
   private observer: MutationObserver;
 
   @Listen("click") onClick(e) {
     // prevent duplicate click events that occur
     // when the component is wrapped in a label and checkbox is clicked
-
     if (
       (this.el.closest("label") && e.target === this.inputProxy) ||
       (!this.el.closest("label") && e.target === this.el)
     ) {
-      this.switched = !this.switched;
+      this.updateSwitch(e);
     }
   }
 
   @Listen("keydown") keyDownHandler(e: KeyboardEvent) {
     if (e.keyCode === SPACE || e.keyCode === ENTER) {
-      e.preventDefault();
-      this.switched = !this.switched;
+      this.updateSwitch(e);
     }
   }
 
   @Watch("switched") switchWatcher() {
-    this.calciteSwitchChange.emit();
     this.switched
       ? this.inputProxy.setAttribute("checked", "")
       : this.inputProxy.removeAttribute("checked");
@@ -78,10 +75,6 @@ export class CalciteSwitch {
 
     let scale = ["s", "m", "l"];
     if (!scale.includes(this.scale)) this.scale = "m";
-
-    let theme = ["dark", "light"];
-    if (!theme.includes(this.theme)) this.theme = "light";
-
     this.setupProxyInput();
   }
 
@@ -94,9 +87,12 @@ export class CalciteSwitch {
   }
 
   render() {
-    const dir = getElementDir(this.el);
     return (
-      <Host role="checkbox" dir={dir} aria-checked={this.switched.toString()} tabIndex={this.tabIndex}>
+      <Host
+        role="checkbox"
+        aria-checked={this.switched.toString()}
+        tabIndex={this.tabIndex}
+      >
         <div class="track">
           <div class="handle" />
         </div>
@@ -146,4 +142,10 @@ export class CalciteSwitch {
     this.inputProxy.setAttribute("name", this.name);
     this.inputProxy.setAttribute("value", this.value);
   };
+  private updateSwitch(e) {
+    e.preventDefault();
+    this.switched = !this.switched;
+    this.change.emit();
+    this.calciteSwitchChange.emit();
+  }
 }