+++ src/components/calcite-tooltip/calcite-tooltip.e2e.ts
@@ -3,7 +3,10 @@ import { newE2EPage } from "@stencil/core/testing";
 import { defaults, hidden, renders } from "../../tests/commonTests";
 
 describe("calcite-tooltip", () => {
-  it("renders", async () => renders("calcite-tooltip"));
+  it("renders", async () =>
+    renders(
+      `<calcite-tooltip open reference-element="ref"></calcite-tooltip><div id="ref">😄</div>`
+    ));
 
   it("honors hidden attribute", async () => hidden("calcite-tooltip"));
 
@@ -28,10 +31,6 @@ describe("calcite-tooltip", () => {
       {
         propertyName: "referenceElement",
         defaultValue: undefined
-      },
-      {
-        propertyName: "theme",
-        defaultValue: "light"
       }
     ]));
 
@@ -61,7 +60,7 @@ describe("calcite-tooltip", () => {
     const page = await newE2EPage();
 
     await page.setContent(
-      `<style>.hydrated--invisible {visibility: hidden;}</style><calcite-tooltip></calcite-tooltip><div>referenceElement</div>`
+      `<calcite-tooltip></calcite-tooltip><div>referenceElement</div>`
     );
 
     const element = await page.find("calcite-tooltip");
@@ -89,7 +88,7 @@ describe("calcite-tooltip", () => {
     const page = await newE2EPage();
 
     await page.setContent(
-      `<style>.hydrated--invisible {visibility: hidden;}</style><calcite-tooltip reference-element="ref" open>content</calcite-tooltip><div id="ref">referenceElement</div>`
+      `<calcite-tooltip reference-element="ref" open>content</calcite-tooltip><div id="ref">referenceElement</div>`
     );
 
     await page.waitForChanges();
@@ -111,7 +110,7 @@ describe("calcite-tooltip", () => {
     const page = await newE2EPage();
 
     await page.setContent(
-      `<style>.hydrated--invisible {visibility: hidden;}</style><calcite-tooltip reference-element="ref">content</calcite-tooltip><div id="ref">referenceElement</div>`
+      `<calcite-tooltip reference-element="ref">content</calcite-tooltip><div id="ref">referenceElement</div>`
     );
 
     await page.waitForChanges();
@@ -131,7 +130,7 @@ describe("calcite-tooltip", () => {
     const page = await newE2EPage();
 
     await page.setContent(
-      `<style>.hydrated--invisible {visibility: hidden;}</style><calcite-tooltip reference-element="ref" open>hi</calcite-tooltip><div id="ref">referenceElement</div>`
+      `<calcite-tooltip reference-element="ref" open>hi</calcite-tooltip><div id="ref">referenceElement</div>`
     );
 
     await page.waitForChanges();