+++ CHANGELOG.md
@@ -5,8 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-
-## Unreleased
+## [Unreleased]
 
 ### Added
 - new `calcite-label`, `calcite-input`, and `calcite-input-message` components
@@ -15,14 +14,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Breaking Changes
 - `calcite-dropdown` - `alignment` attribute now uses `start` and `end` values instead of `left` and `right`
 - `calcite-dropdown-item` - `link-title` attribute has been removed
+- `calcite-tree` - `size` prop is now `scale` to be consistent with other components
 
 ### Updated
 - `calcite-dropdown` - active state indicators for `selection-mode=none` have been removed
 - `calcite-dropdown` - active state indicators for `selection-mode=multi` have been updated to use checkmarks
 - `calcite-dropdown-item` - any attributes passed to a `calcite-dropdown-item` that has a `href` attribute will now be spread to the rendered child link
+- `calcite-loader` - now displays as circle, added fade out at the end of determinate loader
 
 ### Fixed
 - `calcite-dropdown` - `alignment=center` now correctly positions the dropdown if the slotted `dropdown-trigger` is wider than the dropdown container
+- `calcite-loader` - fixed in ie11
+- `calcite-radio-group` - fixed in ie11
+- `calcite-progress` - fixed in ie11
+- `calcite-modal` - fixed in ie11
+- `calcite-tree-item` - fixed in ie11
+- `calcite-tree` - fixed in ie11
 
 ## [v1.0.0-beta.20] - Feb 25th 2020